requirements.txt
aiohttp==3.9.5
aiohttp_client_cache==0.11.0
beautifulsoup4==4.12.3
folium==0.16.0
langchain_core==0.1.52
langchain_groq==0.1.3
motor==3.4.0
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
plotly==5.22.0
pymongo==4.7.1
python-dotenv==1.0.1
spacy==3.7.4
streamlit
tenacity==8.2.3
throttler==1.2.2
spacy
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl
streamlit_folium
lxml
